agents/requirements.txt
google-adk
fastapi
websockets
python-dotenv
google-generativeai